test(a2a): house the unknown-target tests in the mirrored test file
CI caught what 419 local passes did not: PS-204 orphan-test-file. The
project-structure audit mirrors `tests/…/test__X.py` onto
`src/…/_X.py`, and `test__channel_send_unknown_target.py` had no such
module — the code under test lives in `_channel_send_errors.py`, whose
test file already exists.

So the tests move into `test__channel_send_errors.py`, which is where
the convention says tests for that module belong. No test content
changed; the merged file imports the three new names alongside the
existing ones rather than carrying a second mid-file import block.

WORTH RECORDING, because I nearly trusted the wrong signal: the local
suite passed 419/419 on BOTH 3.11 and 3.12 while CI was red. No amount
of pytest would have caught this — it is a repo-structure rule, not a
behaviour. "Local green" and "CI green" are different claims and only
one of them gates merge.

The other audit finding, PS-103 on `uv.lock`, is PRE-EXISTING and NOT
from this branch: the same audit run against the untouched main
checkout reports it too (alongside an untracked HANDOVER file that only
exists there). It is left alone rather than silently whitelisted —
adding a root-whitelist entry to make my own PR green would be exactly
the "configure the check so it passes" move the constitution names.

Verified after the move: _mcp suite 419 passed / 2 skipped on py3.11,
and `audit-all` drops from 2 errors to 1 — the remaining one being the
pre-existing uv.lock finding.
